Aca van varios cambios y modificaciones. Muestra la grilla de compras tanto. Corregi el IdUsuario que se envia en la compra de productos. Tambien corregi algunas validaciones a la hora de Loguearse

+++ Ecommerce/dominio/Usuario.cs
@@ -14,7 +14,7 @@ namespace dominio
     }
     public class Usuario
     {
-        public Int64 Id { get; set; }
+        public int Id { get; set; }
 
         public string Nombre { get; set; }
